Fold SilkNet's GumService into the shared partial class (#3608)
SilkNet's standalone game-host GumService is replaced by the shared partial class GumService (MonoGameGum/GumService.cs, file-linked) plus a SILK arm (Runtimes/SilkNetGum/GumService.Silk.cs) holding the SKCanvas/IInputContext Initialize family, Silk input factories, the double-typed Update, HandleResize, GetWindowSize, and the concrete Default (no [Obsolete] back-compat subclass, so GumServiceCompat.cs is not linked).

Shared-file changes: added an #elif SILK using SkiaGum; arm; a new partial void AssignClipboard() seam replacing the hard-coded MonoGameGumClipboard assignment (that type isn't linked into SilkNet); and relocated CreateSpriteRenderable + the ContentLoader property into the per-platform partials (SkiaGum lacks Sprite.CreateForCurrentPlatform and a concrete ContentLoader). XnaLike/Raylib gained the relocated members + AssignClipboard impl (behavior-identical).

SkiaGum's CustomSetPropertyOnRenderable is made partial so a SilkNet-local partial supplies the LocalizationService/LocalizationServiceChanged/TryGetLocalizationKey holder the shared GumService wires but SkiaGum's copy never grew, keeping SkiaGum/.Wpf/.Maui/.Standalone unaffected. SilkNet's project-load is upgraded to route through the shared FinishInitialize tail.

Tests: SilkNetGum.Tests 12/12; regressions MonoGameGum.Tests 1840/1840, RaylibGum.Tests 467/467, SkiaGum.Tests 206/206. Builds clean: SilkNetGum, KniGum, RaylibGum, SkiaGum, SkiaGum.Wpf.
